+++ x-pack/plugins/cloud_security_posture/public/pages/compliance_dashboard/compliance_charts/risks_table.test.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { getFormattedNum, RisksTable, RisksTableProps } from './risks_table';
+import { RisksTable, RisksTableProps } from './risks_table';
 
 const mockData: RisksTableProps['data'] = [
   {
@@ -24,35 +24,9 @@ const mockData: RisksTableProps['data'] = [
   },
 ];
 
-const mockNums: number[] = [
-  -1000000000, -1000, 0, 100, 1000, 12345, 123456, 1234567, 12345678, 123456789, 1234567890,
-  1234567890000,
-];
-
 describe('<RisksTable />', () => {
   it('renders loading state', () => {
     render(<RisksTable data={mockData} />);
     // expect(screen.getByTestId(CHART_PANEL_TEST_SUBJECTS.LOADING)).toBeInTheDocument();
-  });
-});
-
-describe('getFormattedNum', () => {
-  it('returns correct abbreviation', () => {
-    // tests that the used properties for Intl.NumberFormat are correct
-    const formattedResults = mockNums.map((n) => getFormattedNum(n));
-    expect(formattedResults).toEqual([
-      '-1B',
-      '-1K',
-      '0',
-      '100',
-      '1K',
-      '12.3K',
-      '123.5K',
-      '1.2M',
-      '12.3M',
-      '123.5M',
-      '1.2B',
-      '1.2T',
-    ]);
   });
 });
